Move update functions tu updaters
