+++ AI_NETCORE_API/Domain/Providers/SellOffers/Abstract/ISellOfferProvider.cs
@@ -9,6 +9,5 @@ namespace Domain.Providers.SellOffers.Abstract
     public interface ISellOfferProvider
     {
         IGetSellOffersByUserIdResponse GetSellOffersByUserId(IGetSellOffersByUserIdRequest getSellOffersByUserIdRequest);
-        IWithdrawSellOfferByIdResponse WithdrawSellOfferById(IWithdrawSellOfferByIdRequest withdrawSellOfferByIdRequest);
     }
 }
